Refactor SCF
