+++ src/DFTK.jl
@@ -54,8 +54,9 @@ include("core/Preconditioner.jl")
 export lobpcg
 include("core/lobpcg.jl")
 
-include("core/scf_damped.jl")
-include("core/scf_nlsolve.jl")
+# include("core/scf_damped.jl")
+# include("core/scf_nlsolve.jl")
+include("core/scf.jl")
 
 export energy_ewald
 include("core/energy_ewald.jl")
